feat: add AssemblyAI debug logging and improve file upload handling
- Add debug logging to save AssemblyAI responses with metadata for troubleshooting
- Improve file upload progress tracking and error handling
- Add time parsing utilities and unit tests
- Update help center page with better documentation
- Enhance file dropzone and upload hooks

+++ backend/src/pipeline/interfaces/transcript_processor.py
@@ -11,8 +11,8 @@ from dataclasses import dataclass
 class TranscriptUtterance:
     """Canonical transcript utterance format."""
     speaker_id: str
-    start_seconds: int
-    end_seconds: int
+    start_seconds: float
+    end_seconds: float
     text: str
 
 
@@ -20,11 +20,11 @@ class TranscriptUtterance:
 class TranscriptSection:
     """A section of transcript utterances."""
     utterances: List[TranscriptUtterance]
-    start_time: int
-    end_time: int
+    start_time: float
+    end_time: float
     
     @property
-    def duration_seconds(self) -> int:
+    def duration_seconds(self) -> float:
         return self.end_time - self.start_time
     
     @property
